Fix type in `get_spacegroup_type`

+++ src/FFI.jl
@@ -111,9 +111,9 @@ function get_dataset(cell::Cell; symprec::Real = 1e-8)
     return dataset
 end # function get_dataset
 
-function get_spacegroup_type(hall_number::Int)
+function get_spacegroup_type(hall_number::Integer)
     spgtype = ccall(
-        (:spg_get_spacegroup_type, spglib), SpacegroupType, (Cint,), hall_number
+        (:spg_get_spacegroup_type, spglib), SpacegroupType, (Cint,), Int32(hall_number)
     )
     return spgtype
 end # function get_spacegroup_type
